+++ conf/init.properties_Windows_Firefox
@@ -2,5 +2,6 @@ DRILL_HOST=http://localhost:8047
 DRIVER_TYPE=FIREFOX
 WEBDRIVERS_PATH=webdrivers/geckodriver_win64.exe
 SECURE_DRILL=true
+DEFAULT_TIMEOUT=2
 DRILL_USER_NAME=mapr
 DRILL_USER_PASSWORD=mapr
