+++ src/test/java/org/apache/drillui/test/framework/testng/BaseTest.java
@@ -1,0 +1,12 @@
+package org.apache.drillui.test.framework.testng;
+
+import org.testng.annotations.AfterSuite;
+import org.apache.drillui.test.framework.steps.BaseSteps;
+
+public class BaseTest {
+
+  @AfterSuite
+  public void tearDown() {
+    BaseSteps.tearDown();
+  }
+}
