--- src/test/java/testng/BaseTest.java
@@ -1,12 +1,0 @@
-package testng;
-
-import org.testng.annotations.AfterSuite;
-import steps.BaseSteps;
-
-public class BaseTest {
-
-  @AfterSuite
-  public void tearDown() {
-    BaseSteps.tearDown();
-  }
-}
