--- src/test/java/testng/unsecure/nav/NavigationTest.java
@@ -1,30 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to you under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package testng.unsecure.nav;
-
-import org.testng.annotations.Test;
-import testng.unsecure.BaseUnsecureTest;
-
-import static org.testng.Assert.*;
-
-public class NavigationTest extends BaseUnsecureTest {
-
-  @Test(groups = {"functional"})
-  public void testNav() {
-  }
-
-}
